add exist check for index and column

+++ Database/TeleoptiCCC7/Releases/00000792.sql
@@ -5,18 +5,22 @@
 ----------------  
 SET NOCOUNT ON
 
-ALTER TABLE dbo.FavoriteSearch ADD
-	WfmArea int NOT NULL DEFAULT (1)	
-GO
-		
-DROP INDEX [IX_FavoriteSearch_Creator_Name] ON [dbo].[FavoriteSearch]
+IF EXISTS (SELECT * FROM sys.indexes WHERE name=N'IX_FavoriteSearch_Creator_Name' AND object_id = OBJECT_ID(N'dbo.FavoriteSearch'))
+	DROP INDEX [IX_FavoriteSearch_Creator_Name] ON [dbo].[FavoriteSearch]
 GO
 
-CREATE UNIQUE NONCLUSTERED INDEX [IX_FavoriteSearch_Creator_Name_WfmArea] ON [dbo].[FavoriteSearch]
-(
-	[Name] ASC,
-	[Creator] ASC,
-	[WfmArea] ASC
-)WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, SORT_IN_TEMPDB = OFF, IGNORE_DUP_KEY = OFF, DROP_EXISTING = OFF, ONLINE = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON)
+IF NOT EXISTS (SELECT * FROM sys.columns WHERE Name = N'WfmArea'  
+            AND Object_ID = Object_ID(N'dbo.FavoriteSearch'))
+BEGIN
+	ALTER TABLE dbo.FavoriteSearch ADD
+		WfmArea int NOT NULL DEFAULT (1)	
+
+	CREATE UNIQUE NONCLUSTERED INDEX [IX_FavoriteSearch_Creator_Name_WfmArea] ON [dbo].[FavoriteSearch]
+	(
+		[Name] ASC,
+		[Creator] ASC,
+		[WfmArea] ASC
+	)WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, SORT_IN_TEMPDB = OFF, IGNORE_DUP_KEY = OFF, DROP_EXISTING = OFF, ONLINE = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON)
+END
 GO
